Swift: rename `log` package to `logging`
