+++ swift/extractor/trap/BUILD.bazel
@@ -52,7 +52,7 @@ swift_cc_library(
     visibility = ["//visibility:public"],
     deps = [
         "//swift/extractor/infra/file",
-        "//swift/log",
+        "//swift/logging",
         "@absl//absl/numeric:bits",
     ],
 )
