+++ swift/logging/BUILD.bazel
@@ -1,5 +1,5 @@
 cc_library(
-    name = "log",
+    name = "logging",
     srcs = glob(["*.cpp"]),
     hdrs = glob(["*.h"]),
     visibility = ["//visibility:public"],
