+++ swift/logging/SwiftDiagnostics.h
