+++ swift/logging/SwiftLogging.cpp
@@ -1,4 +1,4 @@
-#include "swift/log/SwiftLogging.h"
+#include "swift/logging/SwiftLogging.h"
 
 #include <filesystem>
 #include <stdlib.h>
